+++ Presentation/OpenTgResearcherConsole/Helpers/TgMenuHelperClientConAdvanced.cs
@@ -2,6 +2,7 @@
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 // ReSharper disable InconsistentNaming
 
+
 namespace OpenTgResearcherConsole.Helpers;
 
 internal partial class TgMenuHelper